requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
Pillow==9.2.0
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.4
django-allauth==0.51.0